Time: 129 ms (80.54%), Space: 111.5 MB (38.32%) - LeetHub

+++ 0295-find-median-from-data-stream/0295-find-median-from-data-stream.java
@@ -14,21 +14,18 @@ class MedianFinder {
         else
             minHeap.offer(num);
 
-         if (maxHeap.size() > minHeap.size() + 1)
+        if (maxHeap.size() > minHeap.size() + 1)
             minHeap.offer(maxHeap.poll());
         if (minHeap.size() > maxHeap.size())
             maxHeap.offer(minHeap.poll());
     }
 
     public double findMedian() {
-        
-       
-
 
         if (maxHeap.size() == minHeap.size()) {
-            return (maxHeap.peek() + minHeap.peek()) / 2.0;
+            return ((double) maxHeap.peek() + minHeap.peek()) / 2.0;
         }
-            return maxHeap.peek();
+        return maxHeap.peek();
     }
 }
